move all the backfill related functions to another script
